+++ button.h
@@ -17,7 +17,7 @@ class Button
 		 * @param text The text which the button displays
 		 * @note the width and height are applied automatically
 		 */		
-		Button(MI0283QT9 lcd, int x, int y, char* text);
+		Button(MI0283QT9 lcd, uint16_t x, uint8_t y, char* text);
 		
 		/**
 		 * Constructs a Button and centers it
@@ -26,9 +26,8 @@ class Button
 		 * @param y Begin y coordinate
 		 * @param text The text which the button displays
 		 * @note the width and height are applied automatically
-		 */	
-		
-		Button(MI0283QT9 lcd, int y, char* text);
+		 */			
+		Button(MI0283QT9 lcd, uint8_t y, char* text);
 		
 };
